requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-google-genai>=1.0.0
langchain-community>=0.3.0
chromadb>=0.4.18
google-generativeai>=0.3.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
tiktoken>=0.5.2
pytest>=7.4.3
sentence-transformers>=2.2.2
lxml>=4.9.3




